use new DatePicker everywhere
