--- libs/expo/shared/ui-components/src/lib/DatePicker_V2/index.ts
@@ -1,1 +1,0 @@
-export { DatePicker as default } from './DatePicker';
